Move decorator functionality to BaseKubeProcessor where it's used, only
